// Web/Fretter.Web/src/app/core/models/queryClausula.model.ts
export class QueryClausula {
    queryClausulaId: number;

    operadorLogicoId: number;
    descricaoOperadorLogico: string;

    clausulaId: number;
    descricaoClausula: string

    operadorComparacaoId: number;
    descricaoOperadorComparacao: string;

    grupoId: number;
    descricaoGrupo: string;

    valorClausula: string;
    valorClausulaInicial: string;
    valorClausulaFinal: string
}